commit first test

+++ deployment.sh
@@ -2,28 +2,34 @@
 
 # Vérifier si Terraform est installé
 if ! [ -x "$(command -v terraform)" ]; then
-  echo "Installing Terraform..."
+  echo "Installation Terraform..."
   sudo apt install terraform -y
 fi
 
 # initialiser le dossier Terraform
-terraform init
+if ! [ -f .terraform ]; then
+  echo "Initialisation du terraform ..."
+  terraform init
+fi
 
 # appliquer la creation
+echo "Application de la création terraform ..."
 terraform apply
 
 # Vérifier si Ansible est installé
 if ! [ -x "$(command -v ansible)" ]; then
-  echo "Installing Ansible..."
+  echo "Installation Ansible..."
   sudo apt install ansible -y 
 fi
 
 # Generer les fichiers hosts
-echo "Generer le fichier host"
-echo "[wordpress]" > hosts
-echo "${terraform output wp_ip} ansible_user=admin" >> hosts
-echo "[db]" >> hosts
-echo "${terraform output db_ip} ansible_user=admin" >> hosts
+if ! [ -f hosts ]; then
+  echo "Generer le fichier host"
+  echo "[wordpress]" > hosts
+  echo "${terraform output wp_ip} ansible_user=admin" >> hosts
+  echo "[db]" >> hosts
+  echo "${terraform output db_ip} ansible_user=admin" >> hosts
+fi
 
 # Appliquer les playbooks Ansible
 ansible-playbook -i hosts wordpress.yml
